pyproject.toml
[tool.poetry]
name = "inspyredgallery"
version = "1.0-dev1"
description = "A photo gallery application"
authors = ["Taylor-Jayde Blackstone <t.blackstone@inspyre.tech>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.12"
streamlit = "^1.21.0"
pillow = "^9.5.0"
torchvision = "^0.15.1"
torch = "^2.0.0"
numpy = "^1.24.2"
scipy = "^1.10.1"
scikit-learn = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
